refactor: funding statusbox 컴포넌트 분리
