+++ src/pages/funding/[id]/index.tsx
@@ -4,10 +4,10 @@ import Header from '@components/common/Header';
 import * as style from './styles';
 import Button from '@components/common/Button';
 import MessageList from '@components/MessageList';
-import StatusTag from '@components/common/StatusTag';
 import DefaultPresentImg from '@assets/images/default_present.svg';
 import Image from 'next/image';
 import { useGetFundingDeatil } from '@hooks/queries/useFunding';
+import FundingStatusBox from '@components/FundingStatusBox';
 
 export default function Funding() {
   const router = useRouter();
@@ -59,24 +59,7 @@ export default function Funding() {
       <article className={style.wrapper}>
         <h2 className={style.title}>{title}</h2>
 
-        <div className={style.statusBox}>
-          <div className={style.subInfoWrapper}>
-            <div className={style.statusMsg}>
-              <span className={style.percentageText}>{percent}%</span>의 마음이 UP 되었어요
-            </div>
-            <StatusTag daysLeft={expirationDate} />
-          </div>
-          <div>progress bar</div>
-
-          <div className={style.captionWrapper}>
-            <span>
-              펀딩 종료까지 <span className={style.blueText}>{expirationDate}일</span>
-            </span>
-            <span>
-              <span className={style.blueText}>목표금액</span> {goalPrice.toLocaleString()}원
-            </span>
-          </div>
-        </div>
+        <FundingStatusBox info={{ percent, expirationDate, goalPrice }} />
 
         <RoleBasedButton />
